test: wait for requests to finish when testing async handling
TestConcurrentAsync has a race condition when writing/reading the 'processed'
field of the batch processor due to the async nature of the processor.
The waitgroup is already handling the concurrent part but because 'handleStream'
returns without waiting for the processor to finish, multiple goroutines can access
the field, leading to flaky tests.

Add a for loop that blocks trying to acquire/wait for the unreleased locks
held by the unfinished processor goroutines.