smoke: Remove usage of artifacts API
Removes the usage of the Elastic artifacts API, in favor of using the
Elastic Cloud API to obtain all the stack pack versions available in the
configured ESS region.

This approach is better than using the artifacts API for two reasons:

    1. Artifacts API and ESS regions may have different stack packs
    2. Unused artifacts are removed after 30 days.

By using the ESS API, we no longer will see failures due to a missing BC
or old versions removed (or the `7.17.x` older versions).

Last, it also adds the `SKIP_DESTROY` conditional to all the `test.sh`
scrits.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>