Move config reloading to beatcmd
beatcmd is now provided with a function that
returns a "Runner", given some parameters:
a complete configuration object containing
both input and output configuration, a logger,
and beat.Info. beatcmd registers with the
libbeat management API for handling config
changes.

beatcmd takes care of stopping and starting
Runners in response to config reloads.

beater is now only responsible for running a
single instance of the APM Server logic with
a static configuration. beater does not
depend on beatcmd by design.

Some of the test helper code in beater has
been moved to a new package, beatertest.