feat: add global labels to service metrics
Add labels to service metrics struct and move comparable fields to
a separate struct.
The new struct is not comparable and cannot be used as a map key so
replace the map with a custom hash map using separate chaining for
collision resolution and implement a custom hash function.

Add systemtest to verify global labels are working as intended.