test: fix TestIntakeAsync/semaphore_size_default flaky test
The test is using the default semaphore size (10) and sending a request
with a lot of events followed by 9 concurrent requests.
The first request (heavy.ndjson) is acquiring and releasing the semaphore
193 times and is also processed asynchronously on the server.
Because of the async nature of the requests, the heavy request can hold
more than a semaphore at the same time leading to a failure (queue is full)
while processing the 9 concurrent requests.
The race condition can be avoided by processing the heavy request synchronously
followed by 10 concurrent request to make sure the async processing on the intake
handler is working.