beater: Set semaphore, modelindexer chan from mem
Update the MaxConcurrentDecoders and the modelindexer internal channel
based on the total amount of memory that the APM server can access. It
looks at cgroups if available, otherwise, falls back to the total system
memory.

Also updates the default semaphore size to `48`, which will be used when
APM server is 1GB or smaller and the default modelindexer internal queue
size to `64` from `100`.

This fixes a couple of issues; 1GB instances could still OOM if enough
concurrency was used (> 80 concurrent agents sending data as fast as
possible), and another issue where APM Server doesn't use the total
available memory to its advantage.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>