Fix wrong punctuation for arguments in curl commands
Removed the redundant symbol " carried in the curl parameter in the example. In the original example, this will cause the id of the sourcemap managed by kibana to have more escape characters when uploading by referring to the example, which will cause the sourcemap to map to the error stack and cannot work. For example The correct id should be "apm-foo-1.0-xxx", the original example would result in an id of "apm:\"demo\"-\"1.0\-xxx"".
This issue is mentioned in [sourcemap to error stack doesn't work](https://github.com/elastic/apm-agent-rum-js/issues/1289)