Simplify Kibana client
Remove the asynchronous connection loop from the
Kibana client, and remove version checking.
We only support Kibana >= APM Server, and all
recent versions of APM Server support the central
config API (the only place we care about the
version.)

Also, remove allowance for (and unit testing of)
an impossible code path related to having a nil
*kibana.Client from the Kibana-based agent config
fetcher. We will now panic if a nil client is
passed into NewKibanaFetcher, enforcing that a
non-nil client is required in the contract.

Extraction of VCS info is moved to package
version, so it can be used for defining the Kibana
client's User-Agent.