modelindexer: Scale active indexers based on load
Modifies the `modelindexer` to scale up and down the number of "active"
indexers (goroutines pulling out of the internal modelindexer queue),
based on consecutive number of flushes up to 20% of the `GOMAXPROCS`.

When N number of full flushes (Reached or exceeded `FlushBytes`) occur,
a new active indexer will be created by the scale action, as long as the
scaling cooldown has elapsed.

Equally, when N number of timed flushes (due to `FlushInterval`) occur,
an active indexer will be scaled down, since not enough load is going
through the server to warrant the current number of active indexers.

Active indexer downscaling can also be triggered due to a change in the
`GOMAXPROCS`. This is particularly important for containerized or cgroup
environments where CPU quotas may be updated "live". In this case, the
downscale cooldown is ignored and active indexers are scaled down until
until `active <= math.RoundToEven(GOMAXPROCS / 4)`.

When traffic isn't going through an active indexer, a new timer has been
introduced to allow completely idle indexers to be scaled back. The idle
check interval can be configured via `IdleInterval`.

Scaling is enabled by default, but can be disabled via the configuration
option `output.elasticsearch.scaling.disabled: true`.

Last, modifies the default settings for `max_requests` and `flush_bytes`
to be `50` and `1MB` respectively. This allows smaller payloads to be
sent to Elasticsearch, more available indexers can be used by instances
with more processing power, and the indexers are cycled faster, which
results in better usage and performance.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>