Stop setting observer.id, observer.ephemeral_id
We don't use these fields in the UI anywhere, and
for debugging we can rely on the hostname to
distinguish one APM Server instance from another.

The observer.id and observer.ephemeral_id fields
suggest that we should care about specific instances,
and their continued presence contradicts our
intentions for a more modern "cloud native" approach,
where APM Server instances may come and go.

For now we need to continue persisting a UUID
to disk for stack monitoring, as this is used to
identify instances of APM Server. In the future
we expect to revisit how stack monitoring works,
and at that point we should redesign the documents
to not enshrine a "pets" mentality.

For tail-based sampling we now generate a UUID
(for identifying its documents) at process start.
The field has been renamed from observer.id to
agent.ephemeral_id, the latter of which is an
ECS field (observer.id is not).