Make ec_deployment observability opt-in again
Replace the var "monitor_deployment" with "observability_deployment",
which is either the ID of an existing deployment, "self" to send
logs and metrics to the same deployment, "dedicated" to create a
dedicated deployment for sending logs and metrics, or "none" to
disable observability.

For ec_deployment, "none" is the default. For testing/cloud, "self"
is the default.