sampling: Discard writes on full badger storage
Discards the badger transaction when the soft limit has reached. This
fixes a problem where each of the sharded readWriters would accumulate
events in the `badger.Tx` until the transaction is full. This led to
ouf of memory issues when the storage soft limit is reached.

Additionally, reduce the number of sharded readWriters to `GOMAXPROCS`
rather than using the `NumCPU()`, since the latter causes containerized
workloads to have (mostly) greater number of sharded readWriters than
it most likely has access to, which will also reduce the memory usage
for TBS.

Last, modifies the options which are used to open the Badger database:

- `Truncate=true`: Allows space reclamationfor vlog unreadable files.
- `NumMemtables=4`: From 5.
- `NumLevelZeroTables=4`: From 5.
- `NumLevelZeroTablesStall=12`: From 15.
- `MaxTableSize=16MB`: From 64MB.
- `ValueLogFileSize=64MB`: From 128MB.

To reduce the overall memory that is consumed by TBS.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>