docker: fix file access for apm-server user (#9604)
* docker: fix file access for apm-server user

Ensure the apm-server.yml is readable by the
apm-server user (UID 1000), and teh data/
directory is writeable by the same user.
This enables users to run the image with a
non-root GID. This was possible in older
versions, as the apm-server user was created
with the root supplementary group.

* Update changelogs/head.asciidoc

Co-authored-by: Marc Lopez Rubio <marc5.12@outlook.com>

* Use COPY --chown

Co-authored-by: Marc Lopez Rubio <marc5.12@outlook.com>
(cherry picked from commit c5141330431742e162ad4f7d4deba5ee138b02c1)