agentcfg: Implement ElasticsearchFetcher
ElasticsearchFetcher is always enabled. It runs a background goroutine
to poll Elasticsearch to refresh a agentcfg cache in memory and serves
fetch requests from the cache. If the query to Elasticsearch is not
successful, it will fallback to the existing behavior, which is to
serve fetch requests using on-demand kibana requests or through fleet
config.

During startup and when the cache is not ready, all agentcfg fetch
requests will be responded with 503. If no fallback is enabled and that
the ES search request fails, server will respond 403 to fetch requests,
which is the same as the existing behavior.