feat: update code to support latest kibana PR
sourcemaps are compressed with zlib and encoded as base64, keep that
in mind when retrieving the content from ES.
The document structure has been updated to reflect the current latest
commit in the kibana PR.
Use the _id to fetch the actual sourcemap since the service fields are
doc-value-only and cannot be used for filtering.
Log an error whenever sync fails.