feat: translate otel log events to error events
Translate otel logs events that represent an error to error events.
Based on the otel spec for recording exceptions, at least one of the
following sets of attributes is required:

exception.type
exception.message

Additionally we want to map an app crash to an error event by checking
the event.name.

Add test to validate the expected APMEvent output.