txmetrics: Add `10m`, `1h` aggregation intervals
Modifies the existing `txmetrics` aggregator to support a list of extra
aggregation intervals, by introducing a new `RollUpIntervals` config
option. The configuration option is not meant to be user facing, but
rather allow us to configure smaller intervals for testing purposes.

Two new intervals are added and published by default, 10 and 60m (1h).
The aggregator will also be flushed when the APM Server is stopped or
restarted.

The `metrics-apm.internal` datastream definition has been duplicated to
allow "normal" metrics to be published to `metrics-apm.internal` and the
time-bucketed transaction metrics will now be published in three new
data streams with a time suffix (`1m`, `10m` and `60m`). The two field
definitions are kept in sync via the `genpackage` tool.

`Event.Duration` ecs field has been added to `metrics-apm.internal` to
allow the metric set documents to contain the aggregated interval they
represent.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>