fix: cast event.duration to long in event_duration pipeline  (#9901)
* fix: cast event.duration to long in event_duration pipeline

The otel span start/end dates are used to compute the event
duration of the APM event, stored as a time.Duration.
The duration is converted to nanoseconds stored in an int64
field, however the values is casted to int in the event_duration
pipeline leading to an overflow if the original 64 bit value
cannot be casted safely to the 32-bit int.

* test: update otlp system-test to validate duration overflow fix

* changelog: add changelog entry

(cherry picked from commit f6872f477fd3a605696e6f4a21dcb647cfabe7bc)

# Conflicts:
#	apmpackage/apm/changelog.yml