spanmetrics: Add `10m` `60m` aggregation intervals
Modifies the `spanmetrics` aggregator to support extra aggregation
intervals by introducing a new `RollUpIntervals` (internal only) config
option. The configuration option is not user facing, but rather allow us
to configure smaller intervals for testing purposes. Also removing the
`aggregation.service_destinations.interval` config option.

Two new intervals are added and published by default, 10 and 60m (1h).
and these are flushed when the APM Server is stopped or restarted.

Introduces 3 `metrics-apm.service_destinations.${interval}m` datastreams
which contain the mapping definitions for transaction metrics with a
series of changes to the apmpackage generator to allow interval
templating and individual ILM policies to be tied to each interval.
A new `Metricset.Interval` field is added to the ingested documents, to
allow them to contain the aggregated interval they represent. The field
format is `${interval}m` or `${interval}s` if < 60s.

Older APM Servers sending data to an Elasticsearch cluster with updated
APM Package, will continue indexing metrics to `metrics-apm.internal`.

Last, there's some light refactoring for the `Start` and `Stop` methods
to a common `baseaggregator` package which is imported by each of the
individual aggregators.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>