perf: reduce aggregation initial memory usage
we are currently allocating a map with a size of maxGroups/maxServices/maxSize.
This is consuming a lot of memory, especially for people that do not come close
to the maximum value.
Use (max/16)+1 as an hint to reduce memory usage. The new value will make sure
we have at most 4 reallocation (table doubling = 2->4->8->16) and that the final
value is at least equal to the maximum.