test: write directly to stderr and avoid leaking goroutine
Writing to the stderrPipeWriter blocks forever.
Reads and Writes on the pipe are matched one to one except when
multiple Reads are needed to consume a single Write.
The data is copied directly from the Write to the corresponding Read;
there is no internal buffering.
As a result the consumeStderr goroutine leaks and run forever.
Because the data has already been written to the log at this point,
we can just write directly to stderr.
Don't change the exit error stderr message to avoid duplicate messages.