Relocate most modelprocessors to apm-data
Removes most of the modelprocessors from the apm-server repo and instead
places them in the `apm-data` repository which can be re-used by other
downstream services.

There's a slight awkwardness on the synchronization between the metrics
defined in the `metrics_internal/fields/*_metrics.yml` and the upstream
`modelprocessor.IsInternalMetricName`. To facilitate checking that all
defined fields are set in the upstream repo, the `internal_metrics.go`
generator has been repurposed as a linter which instructs the user to
update the remote definition should the two go out of sync.

Since these metrics are not updated frequently, the nuisance of this
process shouldn't be experienced very often.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>