Universal Profiling: Keep deprecated methods
We still want these to be part of the protobuf specification and resolve to
stub methods, so that 8.7 HAs reporting to 8.8 servers show a version
mismatch error to the user, suggesting an upgrade, and also gracefully
stop execution.

If these methods are absent, the default gRPC "Unimplemented method" is
shown and the HAs keep running.