ci: run with tests in GitHub Actions (#11046)
Run linting, unit tests, and system tests
with GitHub Actions. Remove these from the
Jenkinsfile.

We also remove microbenchmarks and hey-apm
stages from Jenkins, but without replacement.
We are only making critical bug fixes to 7.17,
so these stages are not necessary; we can run
benchmarks locally or on a dedicated machine
as needed.

Use LF (not CRLF) on Windows for testdata
files (*.ndjson, *.json, *.map).