agentcfg: fix TestRun flakiness (#11146)
The test was intermittently failing due to
there being 2 transactions rather than 1,
which would happen if the Run loop triggered
a second refresh. We reduce the possibility
of this happening by polling for the first
transaction to be flushed, rather than waiting
for a full second and then expecting one.