Revise wait_for_integration ES implementation
Instead of polling Elasticsearch for the presence of index templates,
attempt to create a data stream; this will fail if there is no index
template with a matching index pattern. This decouples apm-server
from the integration package, and will enable apm-server to
alternatively work with the apm-data ES plugin.

Instead of eagerly checking the preconditions (which includes making
these attempts to create a data stream), we now delay the checks to
the first attempt to index/publish events. This means that the
"publish_ready" attribute of the server info API response will always
be false until the first time events are sent to apm-server.