Remove apmpackage
Remove apmpackage from the apm-server repo, as we are moving the
integration package to elastic/integrations.

A couple of things have led to this:
 - We are currently making minimal changes to the integration
   package.
 - We are going to switch to an Elasticsearch plugin (apm-data)
   for installing index templates and ingest pipelines, and use
   dynamic mapping more heavily to make the apm-server code and
   templates more loosely coupled.

The integration package will still exist, but will be "input-only",
meaning that it will not define data streams; it will only be
used to define the apm-server configuration/policy vars. To align
with other packages, we will move the integration package to the
elastic/integrations repo. We're doing it now, rather than waiting
for the switchover to the ES plugin, to make progress on phasing
out some legacy CI infrastructure.

What does this mean for development?

 - We won't be able to atomically make changes to the data stream
   templates and pipelines along with code. This will be the case
   when we are switched over to the plugin, we're just bringing
   that forward.
 - System tests, and automated benchmarks and Elastic Cloud
   deployments will use the integration package bundled with
   Kibana, rather than building and installing the package by
   upload.