feat: add chainguard dockerfile
Use chainguard images for base image and building image.
Remove shell script entrypoint and run apm-server directly. We lose the ability
to run other binaries as the entrypoint but that's fine since there are no other
binaries in the image.
Update apm-server config in the builder image since we do not have posix tools in the
static image.
Drop tini since it is now included in docker.