refactor: add log message for explicit MaxRequest setting
DocumentBufferSize is set based on available memory and a message
is logged.
MaxRequests is set based on available memory if a config value was
not provided but only logs a message if the value was derived from
the available memory.
This is a bit confusing so add an log message for when the config
value is used so the logs are consistent