test: set flush_interval to 1ms for es output config
beatertest is using a flush_bytes setting of 1 to avoid delay when flushing data but the
setting has no effect because it's too small.
As a result TestServerTracingEnabled and TestWrapServer are waiting for the flush interval
to kick in which defaults to 1s.
To fix the issue, set the flush_interval to 1ms to flush immediately and avoid delay.
TestWrapServer test time went from 1.2s to 0.01
TestServerTracingEnabled test time went from 1.2s to 0.01