test: write large batches incrementally to avoid huge test time in TestStorageLimit
To cause a badger error we are writing a batch of 150_000 * num_cpu events
This creates some delay on large systems which can be avoided since lower
batch size might be sufficient.
Incrementally send batches of 150_000 events to reduce test time.
Preliminary test shows 1-2 batches are enough to pass the test, lowering
the test time from 1.7s to 0.1s