wait until wrote
Refactor ingest helper into a separate package that wraps apm-perf
telemetrygen. The new package provides a similar interface as
telemetrygen and adds the RunBlockingWait. This function runs telemetry
generation and poll the ES cluster until expected doc count is reached
or timeout, whichever comes first.

NOTE that RunBlockingWait only waits for docs count in data streams that
have a fixed amount of documents after ingestion, ignoring aggregation
related data stream which have different document counts on different
runs.

It updates (esclient.Client).ApmDocCount() to return an easier to assert
on value that is then used in RunBlockingWait. To avoid a circular
dependency telemetrygen does not use the type alias but the underlying
data type.

It replaces assertDocCountGreaterThan and assertDocCountEqual with a
single assertDocCount helper that uses the new data structure and
supports considering previous results in the comparison.