fix: use a single source of truth for the go version
go version is defined in two places: go.mod and .go-version
they often get out of sync and it's not clear which is used where.
CI is using the version defined in go.mod which might be outdated compared
to .go-version.

bump go version in go.mod and ensure scripts are using it consistently

remove old jenkins scripts