fix: label terraform ECH deployments
This deployment does not use any label, but they use the robots org for deployments. The CI that creates the deployments often fails to clean the resources because they are not tagged, we cannot identify the leftovers to clean them. This PR adds the proper tags for all manual terraform deployments that do not use the oblt tools and use the robots organization for deployment.