Improve flaky elasticsearch output test
The previous test was relying on docappender using exactly 1 bulk
indexer to fulfill the request. However, due to the low flush interval
setting of 1ms, there are some occassions where docappender uses 2
instead, resulting in failing unit test.

We could have simply increase the flush interval but I believe that
changing it makes it more consistent in the future.