support patch versions in HasLazyRollover, improve docs
Add support for specifying patch versions in upgrade-config.yaml.
If a patch version is present will take precedence over it's own
minor version, if both are specified.

Patch version use is supported in both keys and values. This allows
to specify, for example, specific patch versions not triggering a
lazy rollover when upgrading to another patch version.

Overhauled the comments in upgrade-config.yaml to reflect this new
possibilities and clarify how to write the configuration with examples.