ci: update changelog release automation (#16358)
* remove old changelog automation

* format release notes for automation

Update release notes format to an easy to manipulate one for
automation.

* add changelog script

* don't create changelog PR, not needed for 9.x

In 9.x release notes are always read from main, so there updating
branch specific release notes is not needed.

* update release docs to detail 8.x releases

* add changelog automation to patch release

The automation can now work reliably on patch releases too, so
run it to reduce the manual intervention from the release manager
on release.

* add empty release notes automatically

Rely on automation to insert release notes when no release note has
been manually added.

This works for:
- breaking changes
- deprecations
- next version new features
- next version new bugfixes

This avoid manual intervention and simplify verifying if a release
contains any of the mentioned changes.