Bump golang.org/x/sync to 0.16.0 to surface panic (#17975)
Action item from #17967

golang.org/x/sync errgroup 0.15.0 does not propagate panics immediately which makes it very hard to detect issues. This is fixed in 0.16.0. Bump dep to 0.16.0.

No changelog as there should be no change in behavior unless there is a panic.

Not feasible to add unit test as panic in goroutine cannot be caught unless there is implementation code change.