feat: return an error if tbs config fails
failure to process the tbs config will log a message and silently
fallback to the default config with tbs disabled.
This is not ideal and doesn't match the behaviour of other configs
which return an error to the caller.

Update the tbs config logic to return an error if processing fails

Update test to assert error on malformed configs