feat: always build multi-arch image for elastic-agent-cloud
To override the cloud image the elastic-agent image is fetched
and modified by replacing the apm-server binary with a local copy.
The GOARCH is hardcoded to amd64 so it's not possible to use a local
apm-server in an arm64 deployment in cloud.

Update make task to always build a multi-platform image and
target linux/amd64 and linux/arm64.

Update Dockerfile to copy the correct binary based on TARGETARCH